LayoutTests/http/tests/inspector/resource-tree/resource-tree-no-xhrs.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    InspectorTest.makeSimpleXHR("GET", "resources/resource.php", false, step2);

    function step2()
    {
        var resource = WebInspector.resourceTreeModel.resourceForURL("http://127.0.0.1:8000/inspector/resource-tree/resources/resource.php");
        InspectorTest.assertTrue(!resource, "XHR resource should not be added to resourceTreeModel.");
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that XHRs are not added to resourceTreeModel.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=60321">Bug 60321</a>
</body>
</html>
