LayoutTests/http/tests/inspector/resource-tree/resource-tree-no-xhrs-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests that XHRs are not added to resourceTreeModel.

Bug 60321
